feat: Car 리스트를 대체하는 1급 컬렉션 Cars 클래스 구현
Cars Car의 List를 포장하는 클래스
RacingGame 구조변경에 맞춰 리팩토링
* 랜덤생성 부분을 Cars로 옮겼음. 이후 인터페이스 분리하여 이를 통해 Cars를 테스트해볼 예정.
